+++ code-types/ExtractAst.js
@@ -1,5 +1,4 @@
 "use strict";
-var TypeConverter = require("./TypeConverter");
 /** Traverse ASTs and extracting type information from method and field signatures
  * @author TeamworkGuy2
  */
@@ -9,15 +8,16 @@ var ExtractAst;
      * Acts recursively, looking up extended types against 'allTypeDefs'. Excludes primitive types from results.
      * @param childTypes the list of 'allTypeDefs' names to extract from
      * @param allTypeDefs a map of all type definition names and CodeAst classes
+     * @param typeFilter a function which filters all the extracted types, could be used to filter out primitive types
      * @param transformTypeName optional transform for the returned type name map
      * @returns a map of type usages found
     */
-    function extractInheritedTypes(childTypes, allTypeDefs, transformTypeName) {
+    function extractInheritedTypes(childTypes, allTypeDefs, typeFilter, transformTypeName) {
         var typesUsed = {};
         for (var i = 0, size = childTypes.length; i < size; i++) {
             var classDef = allTypeDefs[childTypes[i]];
             if (classDef != null) {
-                extractTypesRecursive(classDef, allTypeDefs, true, false, false, false, typesUsed, transformTypeName);
+                extractTypesRecursive(classDef, allTypeDefs, true, false, false, typeFilter, typesUsed, transformTypeName);
             }
         }
         return typesUsed;
@@ -31,7 +31,7 @@ var ExtractAst;
      * @param transformTypeName optional field type transformer, returns a new type string for a given type string, returns null to ignore a field type
      * @returns a map of type usages found
      */
-    function extractAllTypes(childTypes, allTypeDefs, includePrimitiveTypes, transformTypeName, includeInheritedTypes, includeFieldTypes, includeMethodTypes) {
+    function extractAllTypes(childTypes, allTypeDefs, typeFilter, transformTypeName, includeInheritedTypes, includeFieldTypes, includeMethodTypes) {
         if (includeInheritedTypes === void 0) { includeInheritedTypes = true; }
         if (includeFieldTypes === void 0) { includeFieldTypes = true; }
         if (includeMethodTypes === void 0) { includeMethodTypes = true; }
@@ -40,13 +40,13 @@ var ExtractAst;
         for (var i = 0, size = childTypes.length; i < size; i++) {
             var classDef = allTypeDefs[childTypes[i]];
             if (classDef != null) {
-                extractTypesRecursive(classDef, allTypeDefs, includeInheritedTypes, includeFieldTypes, includeMethodTypes, includePrimitiveTypes, typesUsed, transformTypeName);
+                extractTypesRecursive(classDef, allTypeDefs, includeInheritedTypes, includeFieldTypes, includeMethodTypes, typeFilter, typesUsed, transformTypeName);
             }
         }
         return typesUsed;
     }
     ExtractAst.extractAllTypes = extractAllTypes;
-    function extractTypesRecursive(classDef, typeDefs, includeInheritedTypes, includeFieldTypes, includeMethodTypes, includePrimitiveTypes, typesDst, transformTypeName) {
+    function extractTypesRecursive(classDef, typeDefs, includeInheritedTypes, includeFieldTypes, includeMethodTypes, typeFilter, typesDst, transformTypeName) {
         if (classDef == null) {
             return typesDst;
         }
@@ -58,16 +58,14 @@ var ExtractAst;
                 var extendType = extendTypes[m];
                 var typeName = transformTypeName != null ? transformTypeName(extendType) : extendType;
                 var typeNotYetAdded = (typeName == null || typesDst[typeName] == null); // check whether this type has been loaded before the next if-statement loads it
-                if (typeName != null) {
-                    if (includePrimitiveTypes || (!TypeConverter.isPrimitive(typeName) && !TypeConverter.isCore(typeName) && typeName !== "void")) {
-                        var tu1 = typesDst[typeName] || (typesDst[typeName] = []);
-                        tu1.push({ class: classDef, extendType: extendType });
-                    }
+                if (typeName != null && typeFilter(typeName)) {
+                    var tu1 = typesDst[typeName] || (typesDst[typeName] = []);
+                    tu1.push({ class: classDef, extendType: extendType });
                 }
                 var parentClassDef = typeDefs[extendType];
                 // prevent recursive loops, once a type has been loaded once (is not null in the 'typesDst' map), we don't need to check it again
                 if (parentClassDef != null && typeNotYetAdded) {
-                    extractTypesRecursive(parentClassDef, typeDefs, includeInheritedTypes, includeFieldTypes, includeMethodTypes, includePrimitiveTypes, typesDst, transformTypeName);
+                    extractTypesRecursive(parentClassDef, typeDefs, includeInheritedTypes, includeFieldTypes, includeMethodTypes, typeFilter, typesDst, transformTypeName);
                 }
             }
         }
@@ -80,15 +78,13 @@ var ExtractAst;
                     var implementType = implementTypes[m];
                     var typeName = transformTypeName != null ? transformTypeName(implementType) : implementType;
                     var typeNotYetAdded = (typeName == null || typesDst[typeName] == null);
-                    if (typeName != null) {
-                        if (includePrimitiveTypes || (!TypeConverter.isPrimitive(typeName) && !TypeConverter.isCore(typeName) && typeName !== "void")) {
-                            var tu1 = typesDst[typeName] || (typesDst[typeName] = []);
-                            tu1.push({ class: classDef, implementType: implementType });
-                        }
+                    if (typeName != null && typeFilter(typeName)) {
+                        var tu1 = typesDst[typeName] || (typesDst[typeName] = []);
+                        tu1.push({ class: classDef, implementType: implementType });
                     }
                     var parentInterfaceDef = typeDefs[implementType];
                     if (parentInterfaceDef != null && typeNotYetAdded) {
-                        extractTypesRecursive(parentInterfaceDef, typeDefs, includeInheritedTypes, includeFieldTypes, includeMethodTypes, includePrimitiveTypes, typesDst, transformTypeName);
+                        extractTypesRecursive(parentInterfaceDef, typeDefs, includeInheritedTypes, includeFieldTypes, includeMethodTypes, typeFilter, typesDst, transformTypeName);
                     }
                 }
             }
@@ -103,15 +99,13 @@ var ExtractAst;
                 for (var m = 0, sizeM = fieldTypes.length; m < sizeM; m++) {
                     var typeName = transformTypeName != null ? transformTypeName(fieldTypes[m]) : fieldTypes[m];
                     var typeNotYetAdded = (typeName == null || typesDst[typeName] == null);
-                    if (typeName != null) {
-                        if (includePrimitiveTypes || (!TypeConverter.isPrimitive(typeName) && !TypeConverter.isCore(typeName) && typeName !== "void")) {
-                            var tu1 = typesDst[typeName] || (typesDst[typeName] = []);
-                            tu1.push({ class: classDef, field: field });
-                        }
+                    if (typeName != null && typeFilter(typeName)) {
+                        var tu1 = typesDst[typeName] || (typesDst[typeName] = []);
+                        tu1.push({ class: classDef, field: field });
                     }
                     var fieldTypeDef = typeDefs[fieldTypes[m]];
                     if (fieldTypeDef != null && typeNotYetAdded) {
-                        extractTypesRecursive(fieldTypeDef, typeDefs, includeInheritedTypes, includeFieldTypes, includeMethodTypes, includePrimitiveTypes, typesDst, transformTypeName);
+                        extractTypesRecursive(fieldTypeDef, typeDefs, includeInheritedTypes, includeFieldTypes, includeMethodTypes, typeFilter, typesDst, transformTypeName);
                     }
                 }
             }
@@ -126,15 +120,13 @@ var ExtractAst;
                 for (var m = 0, sizeM = returnTypes.length; m < sizeM; m++) {
                     var typeName = transformTypeName != null ? transformTypeName(returnTypes[m]) : returnTypes[m];
                     var typeNotYetAdded = (typeName == null || typesDst[typeName] == null);
-                    if (typeName != null) {
-                        if (includePrimitiveTypes || (!TypeConverter.isPrimitive(typeName) && !TypeConverter.isCore(typeName) && typeName !== "void")) {
-                            var tu1 = typesDst[typeName] || (typesDst[typeName] = []);
-                            tu1.push({ class: classDef, method: method, returnType: method.returnType });
-                        }
+                    if (typeName != null && typeFilter(typeName)) {
+                        var tu1 = typesDst[typeName] || (typesDst[typeName] = []);
+                        tu1.push({ class: classDef, method: method, returnType: method.returnType });
                     }
                     var returnTypeDef = typeDefs[returnTypes[m]];
                     if (returnTypeDef != null && typeNotYetAdded) {
-                        extractTypesRecursive(returnTypeDef, typeDefs, includeInheritedTypes, includeFieldTypes, includeMethodTypes, includePrimitiveTypes, typesDst, transformTypeName);
+                        extractTypesRecursive(returnTypeDef, typeDefs, includeInheritedTypes, includeFieldTypes, includeMethodTypes, typeFilter, typesDst, transformTypeName);
                     }
                 }
                 // extract parameters' type(s)
@@ -145,15 +137,13 @@ var ExtractAst;
                     for (var m = 0, sizeM = parameterTypes.length; m < sizeM; m++) {
                         var typeName = transformTypeName != null ? transformTypeName(parameterTypes[m]) : parameterTypes[m];
                         var typeNotYetAdded = (typeName == null || typesDst[typeName] == null);
-                        if (typeName != null) {
-                            if (includePrimitiveTypes || (!TypeConverter.isPrimitive(typeName) && !TypeConverter.isCore(typeName) && typeName !== "void")) {
-                                var tu1 = typesDst[typeName] || (typesDst[typeName] = []);
-                                tu1.push({ class: classDef, method: method, parameter: parameter });
-                            }
+                        if (typeName != null && typeFilter(typeName)) {
+                            var tu1 = typesDst[typeName] || (typesDst[typeName] = []);
+                            tu1.push({ class: classDef, method: method, parameter: parameter });
                         }
                         var parameterTypeDef = typeDefs[parameterTypes[m]];
                         if (parameterTypeDef != null && typeNotYetAdded) {
-                            extractTypesRecursive(parameterTypeDef, typeDefs, includeInheritedTypes, includeFieldTypes, includeMethodTypes, includePrimitiveTypes, typesDst, transformTypeName);
+                            extractTypesRecursive(parameterTypeDef, typeDefs, includeInheritedTypes, includeFieldTypes, includeMethodTypes, typeFilter, typesDst, transformTypeName);
                         }
                     }
                 }
@@ -164,11 +154,11 @@ var ExtractAst;
     /** Get a map of the types extracted from annotation arguments from the specified subset of class definitions
      * @param childTypes the list of 'typeDefs' names to extract annotation argument types from
      * @param typeDefs a map of all type definition names to their CodeAst's
-     * @param includePrimitiveTypes whether to include primitive field types (number, boolean, string, etc) in the returned map
+     * @param typeFilter filter for all extracted annotation arguments, for example could filter out primitive field types (number, boolean, string, etc) in the returned map
      * @param extractAnnotationArgument the annotation argument type extractor, returns a string if the annotation argument is successfully mapped to a type, returns null to ignore an annotation argument
      * @returns a map of types found
      */
-    function extractAnnotationArgumentTypes(childTypes, typeDefs, includePrimitiveTypes, transformAnnotationArgument) {
+    function extractAnnotationArgumentTypes(childTypes, typeDefs, typeFilter, transformAnnotationArgument) {
         var typesUsed = {};
         // Loop through all used types to find any missing children
         for (var i = 0, size = childTypes.length; i < size; i++) {
@@ -180,11 +170,9 @@ var ExtractAst;
                     var argNames = Object.keys(annotation.arguments);
                     for (var m = 0, sizeM = argNames.length; m < sizeM; m++) {
                         var argType = transformAnnotationArgument(annotation, argNames[m], annotation.arguments[argNames[m]]);
-                        if (argType != null) {
-                            if (includePrimitiveTypes || (!TypeConverter.isPrimitive(argType) && !TypeConverter.isCore(argType) && argType !== "void")) {
-                                var tu = typesUsed[argType] = typesUsed[argType] || [];
-                                tu.push({ class: classDef, annotation: annotation });
-                            }
+                        if (argType != null && typeFilter(argType)) {
+                            var tu = typesUsed[argType] = typesUsed[argType] || [];
+                            tu.push({ class: classDef, annotation: annotation });
                         }
                     }
                 }