+++ test/strings/StringArrayTest.ts
@@ -4,9 +4,9 @@ import StringArray = require("../../strings/StringArray");
 
 var asr = chai.assert;
 
-suite("StringArray", function StringArrayTest() {
+suite("StringArray", function () {
 
-    test("new StringArray", function StringArrayInstTest() {
+    test("new StringArray()", function () {
         var inst = new StringArray();
         var res = StringArray.of("a")
             .addAll(["", "b", 123])
@@ -22,6 +22,30 @@ suite("StringArray", function StringArrayTest() {
             "\t{\n" +
             "\n" +
             "\t}");
+    });
+
+
+    test("flattenMapJoin", function () {
+        var res = StringArray.flattenMapJoin({
+            alpha: ["alpha"],
+            count: ["1", "2", "3", "4", "5"],
+            last: [""],
+        }, ["-"]);
+
+        asr.deepEqual(res, [
+            "alpha", "-", "1", "2", "3", "4", "5", "-", ""
+        ]);
+    });
+
+
+    test("", function () {
+        var prefix = "+";
+        var suffix = ":";
+        var res = StringArray.preAppend(prefix, suffix, ["alpha", "beta", "end"]);
+
+        asr.deepEqual(res, [
+            "+alpha:", "+beta:", "+end:"
+        ]);
     });
 
 });